server/pyproject.toml
[tool.poetry]
name = "blackjack"
version = "0.1.0"
description = ""
authors = ["Şahin Akkaya <sahin@sahinakkaya.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
requests = "^2.32.3"
transitions = "^0.9.2"
python-socketio = {extras = ["asyncio"], version = "^5.11.4"}
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
